+++ deployment/run-unit-tests.sh
@@ -12,27 +12,30 @@ source_dir="$PWD/../source"
 
 echo "cd $source_dir/modules/b2.core"
 cd $source_dir/modules/b2.core
-npm install
+npm ci
 # npm test
 
 echo "cd $source_dir/services/core"
 cd $source_dir/services/core
-npm install
+npm ci
 # npm test
 
 echo "cd $source_dir/services/polly-service"
 cd $source_dir/services/polly-service
-npm install
+npm ci
 # npm test
 
 echo "cd $source_dir/services/train-model"
 cd $source_dir/services/train-model
-npm install
+npm ci
 # npm test
 
 echo "cd $source_dir/samples/bot-weather-forecast"
 cd $source_dir/samples/bot-weather-forecast
-# npm test
+python3 setup.py install
+pip3 install -r requirements.txt --target .
+pip3 install -r dev-requirements.txt
+python3 test.py
 
 echo "cd $source_dir/samples/leave-feedback"
 cd $source_dir/samples/leave-feedback